Dungeon Generator Rework
Found out I'll be needing to have more control on how my dungeon generates so got rid of the Invoke call method and am restructuring it as a list for the dungeon to look at and go through. Overall got it to generate a few hallways and created rooms out of them. Should be easy to add in a bit more functionality and collision detection in the near future.